requirements.txt
pandas~=1.3.4
requests~=2.26.0
openpyxl~=3.0.10
beautifulsoup4~=4.10.0
selenium~=4.1.0
python-dateutil~=2.8.2
pybit~=1.3.2
Bybit~=0.2.12
numpy~=1.21.4